+++ third-party/line-4.5.2/test/fixtures/cookbooks/test/recipes/filter_lines_between.rb
@@ -1,0 +1,14 @@
+#
+# Verify the results of using the between filter
+#
+
+template '/tmp/between' do
+  source 'samplefile3.erb'
+  action :create_if_missing
+end
+
+filter_lines 'Change lines between matches' do
+  path '/tmp/between'
+  sensitive false
+  filters between: [/^empty/, /last_list/, ['add line']]
+end