+++ third-party/line-4.5.2/test/fixtures/cookbooks/test/recipes/filter_lines_delete_between.rb
@@ -1,0 +1,14 @@
+#
+# Verify the results of using the delete_between filter
+#
+
+template '/tmp/delete_between' do
+  source 'samplefile3.erb'
+  action :create_if_missing
+end
+
+filter_lines 'Delete lines between matches' do
+  path '/tmp/delete_between'
+  sensitive false
+  filters delete_between: [/^empty_list/, /^list/, /kernel/]
+end