+++ third-party/selinux-6.0.4/test/cookbooks/selinux_test/recipes/module_remove.rb
@@ -1,0 +1,3 @@
+selinux_module 'test' do
+  action %i(remove delete)
+end